+++ src_docs/components/date-range.component.jsx
@@ -19,6 +19,7 @@ export default class DateRangeView extends React.PureComponent {
           <DateRange
             id="example3"
             enabled={{ absolute: true, relative: true, period: true }}
+            absoluteRange={{ className: 'test', dateFormat: 'L', numberOfMonths: 2 }}
           />
         </div>
       </React.Fragment>